// types/echo.d.ts
export type QuestionItem = {
  id: string;
  sectionName: string;
  question: string;
  isRequired: string;
  answerType: string | null;
  options: string | null;
  eventAddress: string;
  ticketId: number;
  createdAt: string;
  updatedAt: string;
};

type SavedEvent = {
  accountAddress: string;
};

type DiscountType = "PERCENT" | "FIXED";

export interface TicketDiscount {
  utilityLeft: number;
  isEnabled: boolean;
  limitPerUser: number;
  ticketType: number;
  discountValue: number;
  discountType: DiscountType;
  code: string;
  expireAt: string;
}

export type TicketDiscountList = TicketDiscount[];

export type IEventType = {
  id: string;
  eventAddress: string;
  eventName: string;
  createdAt: string;
  startAt: string;
  endAt: string;
  eventImg: string;
  organiser: string;
  tags: string[];
  category: string;
  hosts: string;
  location: string;
  ticketTypeCount: number;
  ticketsCount: number;
  creator: string;
  minPrice: number;
  maxPrice: number;
  eventType: number;
  currency: string;
  eventLink: string;
  savedEvents: SavedEvent[];
};

export type QuestionList = QuestionItem[];